Enable colors for ColoredTriangle example

+++ SilkNetOpenGL/SilkNetTriangle/Program.cs
@@ -67,7 +67,7 @@ namespace SilkNetTriangle
 
         private static unsafe void DrawTriangle() 
         {
-            double[] vertices = { 
+            float[] vertices = { 
                 -1.0f, -1.0f, 0.0f,
                 1.0f, -1.0f, 0.0f,
                 0.0f, 1.0f, 0.0f,
